add block_signals tests

+++ diff/workers/sqs.py
@@ -1,11 +1,10 @@
-import boto3
 import json
-import traceback
-from diff.diff import diff_snapshot
-from render.snapshot import render_snapshot, finished
+
+import boto3
+
+from render.snapshot import finished
 from utils.send_message import send_message
 from utils.settings import *
-
 from .process_message import process_message
 
 sqs = boto3.resource('sqs')
@@ -23,14 +22,10 @@ print('received messages: {}'.format(len(messages)))
 while len(messages) > 0:
     messages = sorted(messages, key=lambda m: json.loads(m.body, encoding='utf-8')['browser'])
     for message in messages:
-        try:
-            message_data = process_message(message.body)
+        message_data = process_message(message.body)
+        if message_data is not None:
             send_message(message_data)
             message.delete()
-        except Exception as ex:
-            print(''.join(traceback.format_exception(etype=type(ex), value=ex, tb=ex.__traceback__)))
-            print('There was an error trying to render the snapshot {}'.format(title))
-            pass
 
     messages = get_messages()
     print('received more messages: {}'.format(len(messages)))
